merge develop to 1520

+++ backend/unpp_api/apps/management/serializers.py
@@ -3,15 +3,14 @@ from django.db.models import Q
 from rest_framework import serializers
 
 from account.models import User
+from agency.models import AgencyMember, AgencyOffice
 from agency.roles import AgencyRole
 from management.fields import CurrentAgencyFilteredPKField, CurrentPartnerFilteredPKField
-from agency.models import AgencyMember, AgencyOffice
 from management.invites import send_partner_user_invite, send_agency_user_invite
 from partner.models import PartnerMember, Partner
 
 
 class AgencyOfficeManagementSerializer(serializers.ModelSerializer):
-
     class Meta:
         model = AgencyOffice
         fields = (
@@ -21,7 +20,6 @@ class AgencyOfficeManagementSerializer(serializers.ModelSerializer):
 
 
 class AgencyMemberManagementSerializer(serializers.ModelSerializer):
-
     office = AgencyOfficeManagementSerializer(read_only=True)
     office_id = CurrentAgencyFilteredPKField(queryset=AgencyOffice.objects.all(), write_only=True)
     role_display = serializers.CharField(source='get_role_display', read_only=True)
@@ -47,8 +45,8 @@ class AgencyMemberManagementSerializer(serializers.ModelSerializer):
 
 
 class AgencyUserManagementSerializer(serializers.ModelSerializer):
-
     office_memberships = AgencyMemberManagementSerializer(many=True, source='agency_members', allow_empty=False)
+    email = serializers.EmailField()
 
     class Meta:
         model = User
@@ -61,7 +59,11 @@ class AgencyUserManagementSerializer(serializers.ModelSerializer):
             'status',
             'office_memberships',
         )
-        extra_kwargs = {'fullname': {'required': True}}
+        extra_kwargs = {
+            'fullname': {
+                'required': True
+            },
+        }
 
     def validate(self, attrs):
         validated_data = super(AgencyUserManagementSerializer, self).validate(attrs)
@@ -83,6 +85,13 @@ class AgencyUserManagementSerializer(serializers.ModelSerializer):
     @transaction.atomic
     def save(self):
         update = bool(self.instance)
+
+        if not update:
+            existing_user = User.objects.filter(email__iexact=self.validated_data['email']).first()
+            if existing_user and existing_user.is_agency_user:
+                self.instance = existing_user
+                self.partial = True
+
         user = super(AgencyUserManagementSerializer, self).save()
 
         if self.context['agency_members'] is not None:
@@ -93,9 +102,11 @@ class AgencyUserManagementSerializer(serializers.ModelSerializer):
                     office=member_data.pop('office_id'),
                     defaults=member_data
                 )[0].pk)
-            AgencyMember.objects.filter(user=user).exclude(pk__in=memberships).delete()
 
-        if not update:
+            if update:
+                AgencyMember.objects.filter(user=user).exclude(pk__in=memberships).delete()
+
+        if not update and not self.partial:
             user.profile.accepted_tos = True
             user.profile.save()
             user.set_unusable_password()
@@ -106,7 +117,6 @@ class AgencyUserManagementSerializer(serializers.ModelSerializer):
 
 
 class PartnerOfficeManagementSerializer(serializers.ModelSerializer):
-
     name = serializers.CharField(source='get_country_code_display')
 
     class Meta:
@@ -122,14 +132,12 @@ class CurrentUserOfficesListSerializer(serializers.ListSerializer):
     def to_representation(self, data):
         request = self.context.get('request')
         if request:
-            data = data.filter(
-                Q(partner_id=request.active_partner.id) | Q(partner__hq_id=request.active_partner.id)
-            )
+            data = data.filter(partner_id__in=request.user.partner_ids)
+
         return super(CurrentUserOfficesListSerializer, self).to_representation(data)
 
 
 class PartnerMemberManagementSerializer(serializers.ModelSerializer):
-
     office = PartnerOfficeManagementSerializer(read_only=True, source='partner')
     office_id = CurrentPartnerFilteredPKField(queryset=Partner.objects.all(), write_only=True)
     role_display = serializers.CharField(source='get_role_display', read_only=True)
@@ -147,8 +155,8 @@ class PartnerMemberManagementSerializer(serializers.ModelSerializer):
 
 
 class PartnerUserManagementSerializer(serializers.ModelSerializer):
-
     office_memberships = PartnerMemberManagementSerializer(many=True, source='partner_members', allow_empty=False)
+    email = serializers.EmailField()
 
     class Meta:
         model = User
@@ -161,7 +169,11 @@ class PartnerUserManagementSerializer(serializers.ModelSerializer):
             'status',
             'office_memberships',
         )
-        extra_kwargs = {'fullname': {'required': True}}
+        extra_kwargs = {
+            'fullname': {
+                'required': True
+            }
+        }
 
     def validate(self, attrs):
         validated_data = super(PartnerUserManagementSerializer, self).validate(attrs)
@@ -171,22 +183,33 @@ class PartnerUserManagementSerializer(serializers.ModelSerializer):
     @transaction.atomic
     def save(self):
         update = bool(self.instance)
+
+        if not update:
+            existing_user = User.objects.filter(email__iexact=self.validated_data['email']).first()
+            if existing_user and existing_user.is_partner_user:
+                self.instance = existing_user
+                self.partial = True
+
         user = super(PartnerUserManagementSerializer, self).save()
         request = self.context.get('request')
 
         if self.context['partner_members'] is not None:
             memberships = []
             for member_data in self.context['partner_members']:
-                memberships.append(PartnerMember.objects.update_or_create(
-                    user=user,
-                    partner=member_data.pop('office_id'),
-                    defaults=member_data
-                )[0].pk)
-            PartnerMember.objects.filter(user=user).filter(
-                Q(partner_id=request.active_partner.id) | Q(partner__hq_id=request.active_partner.id)
-            ).exclude(pk__in=memberships).delete()
+                partner = member_data.pop('office_id', None)
+                if partner:
+                    memberships.append(PartnerMember.objects.update_or_create(
+                        user=user,
+                        partner=partner,
+                        defaults=member_data
+                    )[0].pk)
 
-        if not update:
+            if update:
+                PartnerMember.objects.filter(user=user).filter(
+                    Q(partner_id=request.active_partner.id) | Q(partner__hq_id=request.active_partner.id)
+                ).exclude(pk__in=memberships).delete()
+
+        if not update and not self.partial:
             user.profile.accepted_tos = True
             user.profile.save()
             user.set_random_password()
